fix: Avoid FPs when Composer product name has php
Previously, only PHP package's vendor (product URL's namespace) was
considered as evidence. As of DependencyCheck v12, specifically change
from b51921f, the product (name) is also being considered as evidence.

This results in new false positives. PHP Composer checks are affected,
for example, considering package `pkg:composer/phpunit/php-timer@6.0.0`
as `cpe:2.3:a:php:php:6.0:*:*:*:*:*:*:*` resulting in 17 CVEs
(including Critical).

This commit adds two new suppression rules: one for php as the prefix,
and one for php as the suffix. Both can be observed in the wild.
Additionally, underscore is sometimes used instead of hyphen, and should
be respected. Furthermore, there is `symfony/polyfill-php83` which adds
number suffix, which should also be suppressed as it currently maps to
the base `cpe:/a:php:php`.